fix: 'Backend is not available...' error message

+++ packages/dashboard-frontend/src/store/GitOauthConfig/index.ts
@@ -21,6 +21,7 @@ import {
   getOAuthProviders,
   getOAuthToken,
 } from '../../services/backend-client/oAuthApi';
+import { selectAsyncIsAuthorized, selectSanityCheckError } from '../SanityCheck/selectors';
 
 export interface State {
   isLoading: boolean;
@@ -68,11 +69,20 @@ export type ActionCreators = {
 export const actionCreators: ActionCreators = {
   requestGitOauthConfig:
     (): AppThunk<KnownAction, Promise<void>> =>
-    async (dispatch): Promise<void> => {
-      await dispatch({
+    async (dispatch, getState): Promise<void> => {
+      dispatch({
         type: Type.REQUEST_GIT_OAUTH_CONFIG,
         check: AUTHORIZED,
       });
+      if (!(await selectAsyncIsAuthorized(getState()))) {
+        const error = selectSanityCheckError(getState());
+        dispatch({
+          type: Type.RECEIVE_GIT_OAUTH_CONFIG_ERROR,
+          error,
+        });
+        throw new Error(error);
+      }
+
       const gitOauth: IGitOauth[] = [];
       try {
         const oAuthProviders = await getOAuthProviders();
@@ -106,11 +116,20 @@ export const actionCreators: ActionCreators = {
 
   revokeOauth:
     (oauthProvider: api.GitOauthProvider): AppThunk<KnownAction, Promise<void>> =>
-    async (dispatch): Promise<void> => {
-      await dispatch({
+    async (dispatch, getState): Promise<void> => {
+      dispatch({
         type: Type.REQUEST_GIT_OAUTH_CONFIG,
         check: AUTHORIZED,
       });
+      if (!(await selectAsyncIsAuthorized(getState()))) {
+        const error = selectSanityCheckError(getState());
+        dispatch({
+          type: Type.RECEIVE_GIT_OAUTH_CONFIG_ERROR,
+          error,
+        });
+        throw new Error(error);
+      }
+
       try {
         await deleteOAuthToken(oauthProvider);
         dispatch({
